Consolidate commands
