+++ src/index.ts
@@ -5,9 +5,31 @@ import * as types from "vscode-languageserver-types"
 
 import { Event, IEvent } from "oni-types"
 
-import * as Commands from "./Commands"
-
 export type DisposeFunction = () => void
+
+/**
+ * API surface area for registering and executing commands
+ */
+export namespace Commands {
+    export type CommandCallback = (args?: any) => any
+    export type CommandEnabledCallback = () => boolean
+
+    export interface ICommand {
+        command: string
+        name: string
+        detail: string
+        enabled?: CommandEnabledCallback
+        messageSuccess?: string
+        messageFail?: string
+        execute: CommandCallback
+    }
+
+    export interface Api {
+        registerCommand(command: ICommand): void
+        unregisterCommand(commandName: string): void
+        executeCommand(name: string, args?: any): boolean | void
+    }
+}
 
 export interface IToken {
     tokenName: string
